requirements.txt
# Core Dependencies
yt-dlp>=2024.3.10
google-generativeai>=0.3.2
moviepy>=1.0.3
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.1

# Additional Dependencies (installed automatically with moviepy)
imageio>=2.31.1
imageio-ffmpeg>=0.4.9
numpy>=1.24.0

# Optional (for Streamlit UI)
streamlit>=1.28.0
